Fixing up all the header examples
- No importing of external doc examples
- All Elastic logos
- Remove all actual links that might send the user to the index page
- Only the first example shows all the popovers
- The Full screen demo has the rest
