+++ src-docs/src/views/header/header_app_menu.js
@@ -45,31 +45,31 @@ export default () => {
       anchorPosition="downRight"
       closePopover={closeMenu}>
       <EuiKeyPadMenu id={keypadId} style={{ width: 288 }}>
-        <EuiKeyPadMenuItem label="Discover" href="">
+        <EuiKeyPadMenuItem label="Discover">
           <EuiIcon type="discoverApp" size="l" />
         </EuiKeyPadMenuItem>
 
-        <EuiKeyPadMenuItem label="Dashboard" href="">
+        <EuiKeyPadMenuItem label="Dashboard">
           <EuiIcon type="dashboardApp" size="l" />
         </EuiKeyPadMenuItem>
 
-        <EuiKeyPadMenuItem label="Dev Tools" href="">
+        <EuiKeyPadMenuItem label="Dev Tools">
           <EuiIcon type="devToolsApp" size="l" />
         </EuiKeyPadMenuItem>
 
-        <EuiKeyPadMenuItem label="Machine Learning" href="">
+        <EuiKeyPadMenuItem label="Machine Learning">
           <EuiIcon type="machineLearningApp" size="l" />
         </EuiKeyPadMenuItem>
 
-        <EuiKeyPadMenuItem label="Graph" href="">
+        <EuiKeyPadMenuItem label="Graph">
           <EuiIcon type="graphApp" size="l" />
         </EuiKeyPadMenuItem>
 
-        <EuiKeyPadMenuItem label="Visualize" href="">
+        <EuiKeyPadMenuItem label="Visualize">
           <EuiIcon type="visualizeApp" size="l" />
         </EuiKeyPadMenuItem>
 
-        <EuiKeyPadMenuItem label="Timelion" href="" betaBadgeLabel="Beta">
+        <EuiKeyPadMenuItem label="Timelion" betaBadgeLabel="Beta">
           <EuiIcon type="timelionApp" size="l" />
         </EuiKeyPadMenuItem>
       </EuiKeyPadMenu>
