+++ src-docs/src/views/header/header_links.js
@@ -12,7 +12,7 @@ export default () => {
   return (
     <EuiHeader>
       <EuiHeaderSectionItem border="right">
-        <EuiHeaderLogo href="">Product</EuiHeaderLogo>
+        <EuiHeaderLogo>Elastic</EuiHeaderLogo>
       </EuiHeaderSectionItem>
 
       <EuiHeaderSectionItem>
