--- src-docs/src/views/header/header_user_menu.js
@@ -1,81 +1,0 @@
-import React, { useState } from 'react';
-
-import {
-  EuiAvatar,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiHeaderSectionItemButton,
-  EuiLink,
-  EuiText,
-  EuiSpacer,
-  EuiPopover,
-} from '../../../../src/components';
-import { htmlIdGenerator } from '../../../../src/services';
-
-export default () => {
-  const id = htmlIdGenerator()();
-  const [isOpen, setIsOpen] = useState(false);
-
-  const onMenuButtonClick = () => {
-    setIsOpen(!isOpen);
-  };
-
-  const closeMenu = () => {
-    setIsOpen(false);
-  };
-
-  const button = (
-    <EuiHeaderSectionItemButton
-      aria-controls={id}
-      aria-expanded={isOpen}
-      aria-haspopup="true"
-      aria-label="Account menu"
-      onClick={onMenuButtonClick}>
-      <EuiAvatar name="John Username" size="s" />
-    </EuiHeaderSectionItemButton>
-  );
-
-  return (
-    <EuiPopover
-      id={id}
-      ownFocus
-      button={button}
-      isOpen={isOpen}
-      anchorPosition="downRight"
-      closePopover={closeMenu}
-      panelPaddingSize="none">
-      <div style={{ width: 320 }}>
-        <EuiFlexGroup
-          gutterSize="m"
-          className="euiHeaderProfile"
-          responsive={false}>
-          <EuiFlexItem grow={false}>
-            <EuiAvatar name="John Username" size="xl" />
-          </EuiFlexItem>
-
-          <EuiFlexItem>
-            <EuiText>
-              <p>John Username</p>
-            </EuiText>
-
-            <EuiSpacer size="m" />
-
-            <EuiFlexGroup>
-              <EuiFlexItem>
-                <EuiFlexGroup justifyContent="spaceBetween">
-                  <EuiFlexItem grow={false}>
-                    <EuiLink href="">Edit profile</EuiLink>
-                  </EuiFlexItem>
-
-                  <EuiFlexItem grow={false}>
-                    <EuiLink href="">Log out</EuiLink>
-                  </EuiFlexItem>
-                </EuiFlexGroup>
-              </EuiFlexItem>
-            </EuiFlexGroup>
-          </EuiFlexItem>
-        </EuiFlexGroup>
-      </div>
-    </EuiPopover>
-  );
-};
